requirements.txt
tk
requests
json
re
pillow
